Move pattern matching to another package
